move testing-implementations into test-scope
